backend/requirements.txt
fastapi 
uvicorn
python-multipart
starlette
pypdf
python-docx 
openai
python-dotenv
ollama
docling
json_repair 
asyncpg  
sqlalchemy 
databases
fpdf2
psycopg2-binary